+++ data_structures/abstract_hash_table.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from typing import TypeVar, Generic, Tuple
 from data_structures.referential_array import ArrayR
-from data_structures.dunder_protected import ProtectAttributes
+from data_structures.sunder_protected import ProtectAttributes
 
 K = TypeVar('K')
 V = TypeVar('V')
