+++ data_structures/abstract_heap.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from typing import Generic, TypeVar, Literal, Iterable
 from abc import abstractmethod, ABC
-from data_structures.dunder_protected import ProtectAttributes
+from data_structures.sunder_protected import ProtectAttributes
 
 T = TypeVar('T')
