+++ data_structures/abstract_set.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from typing import TypeVar, Generic
-from data_structures.dunder_protected import ProtectAttributes
+from data_structures.sunder_protected import ProtectAttributes
 
 T = TypeVar('T')
