+++ data_structures/referential_array.py
@@ -24,7 +24,7 @@ from ctypes import py_object
 from typing import Generic, Union, TypeVar
 from data_structures.abstract_list import List
 from data_structures.abstract_sorted_list import SortedList
-from data_structures.dunder_protected import ProtectAttributes
+from data_structures.sunder_protected import ProtectAttributes
 
 T = TypeVar('T')
